中間モデルを使って手順とカテゴリーを紐つける

+++ app/controllers/public/bookmarks_controller.rb
@@ -1,21 +1,28 @@
 class Public::BookmarksController < ApplicationController
-  
+  before_action :authenticate_user!
+
   def index
     @bookmarks = Bookmark.all
   end 
   
   def create
-    procedure = Procedurs.find(params[:procedure_id])
-    bookmark = current_user.bookmarks.new(procedure_id: procedure.id)
-    bookmark.save
-    redirect_to procedure_path(procedure)
+    @procedure = Procedure.find(params[:procedure_id])
+    bookmark = @procedure.bookmarks.new(user_id: current_user.id)
+    if bookmark.save
+      redirect_to request.referer
+    else
+      redirect_to request.referer
+    end
   end
 
   def destroy
-    procedure = Procedure.find(params[:procedure_id])
-    bookmark = current_user.bookmarks.find_by(procedure_id: procedure.id)
-    bookmark.destroy
-    redirect_to procedure_path(procedure)
+    @procedure = Procedure.find(params[:procedure_id])
+    bookmark = @post.bookmarks.find_by(user_id: current_user.id)
+    if bookmark.present?
+        bookmark.destroy
+        redirect_to request.referer
+    else
+        redirect_to request.referer
+    end
   end
-  
 end
